enabled warning level 4 in msvs2005 gldebug and fixed/silenced all warnings there possibility of new warnings on non-windows platforms

+++ cd_win.c
@@ -334,7 +334,7 @@ void CD_f (void) {
 
 LONG CDAudio_MessageHandler(HWND hWnd, UINT uMsg, WPARAM wParam, LPARAM lParam)
 {
-	if (lParam != wDeviceID)
+	if ((UINT) lParam != wDeviceID)
 		return 1;
 
 	switch (wParam)
